Explicit cast to (uint32_t) for multitape fragnum
Each instance of (uint32_t)fragnum occurs within with a for loop which checks:
    fragnum * MAXIFRAG < foo->indexlen;
where `foo` is some other variable.

MAXIFRAG is MAXCHUNK (from tar/multitape/multitape_internal.h), which is
261120 (or 2^18 - 2^10).  So, in order for fragnum to fit inside (uint32_t),
we need
    0 < foo->indexlen / 266120 < 2^32

tar/multitape/multitape_metadata.h: ->indexlen is stored in the metadata
format as uint64_t, and multitape_metadata.c also calls it a
<64-bit metaindex length>.  (Not helpful for demonstrating the bound, but
potentially worth noting.)

tar/multitape/multitape_metaindex.c: multitape_metaindex_put() is the function
which first writes a value to ->indexlen:
    mdat->indexlen = buflen;
In this function,
    buflen = mind->hindexlen + mind->cindexlen + mind->tindexlen;
where each of those 3 values are <= UINT32_MAX.  Since MAXIFRAG is more than 3,
we know that (buflen/MAXIFRAG) < UINT32_MAX.

For reference, this is very similar to the bound established in
d95c273a313426243b7a03411ab09542bea67334.

+++ tar/multitape/multitape_fsck.c
@@ -121,7 +121,7 @@ deletearchive(STORAGE_D * SD, struct tapemetadata * tmd)
 
 	/* Delete index fragments. */
 	for (fragnum = 0; fragnum * MAXIFRAG < tmd->indexlen; fragnum++) {
-		multitape_metaindex_fragname(hbuf, fragnum, fraghash);
+		multitape_metaindex_fragname(hbuf, (uint32_t)fragnum, fraghash);
 		if (storage_delete_file(SD, 'i', fraghash))
 			goto err0;
 	}
@@ -268,7 +268,8 @@ phase2(uint64_t machinenum, STORAGE_D * SD,
 		for (fragnum = 0;
 		    fragnum * MAXIFRAG < mdatlist[file]->indexlen;
 		    fragnum++) {
-			multitape_metaindex_fragname(hbuf, fragnum, fraghash);
+			multitape_metaindex_fragname(hbuf, (uint32_t)fragnum,
+			    fraghash);
 			if (findinlist(fraghash, flist, nfiles) == nfiles) {
 				fprintf(stdout,
 				    "  Deleting archive"
@@ -307,7 +308,8 @@ nextfile:	;
 		for (fragnum = 0;
 		    fragnum * MAXIFRAG < mdatlist[file]->indexlen;
 		    fragnum++) {
-			multitape_metaindex_fragname(hbuf, fragnum, fraghash);
+			multitape_metaindex_fragname(hbuf, (uint32_t)fragnum,
+			    fraghash);
 			neededvec[findinlist(fraghash, flist, nfiles)] = 1;
 		}
 	}
